bug fixed

+++ nets/RC3D_resnet.py
@@ -178,7 +178,7 @@ class RC3D(nn.Module):
         loss3 = creterion(object_cls_score[e_index2], object_label[e_index2].long())
         #object_bbox_loss
         object_offset = object_offset.reshape(-1, 2)
-        e_index4 = e_index3 * (self.num_classes - 1) + object_label[e_index3] - 1
+        e_index4 = e_index3 * (self.num_classes - 1) + object_label[e_index3].long() - 1
         loss4 = nn.SmoothL1Loss(reduction = 'mean')(object_offset[e_index4], object_bbox_offset[e_index3])
         if math.isnan(loss2.data) or math.isnan(loss4.data):
             print(loss1.data, loss2.data, loss3.data, loss4.data)
